requirements.txt
fastapi
uvicorn
pdf2image
python-multipart
PyPDF2
PyMuPDF
Pillow

